tests/wpt/tests/css/css-view-transitions/navigation/zero-named-elements.html

<!DOCTYPE html>
<title>View transitions: pageswap navigationactivation for replace navigations</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/">
<link rel="author" href="mailto:khushalsagar@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
@view-transition {
  navigation: auto;
}

html {
  view-transition-name: none;
}
</style>
<script>
const expectedUrl = location.href + '?new';

const params = new URLSearchParams(location.search);
// The page the popup navigates to.
const is_new_page = params.has('new');
// The initial page in the popup.
const is_popup_page = params.has('popup');
// The test page itself.
const is_test_page = !is_popup_page && !is_new_page;

const channel = new BroadcastChannel("testchannel");

if (is_test_page) {
  const expectedEvents = ["pageswap", "pagehide", "pagereveal", "finished"];

  promise_test(async t => {
    let popup;
    onload = () => {
      window.events = [];
      popup = window.open("?popup");
      t.add_cleanup(() => popup.close());
    };

    await new Promise(resolve => {
      channel.addEventListener(
        "message", t.step_func(async (e) => {
          if (e.data === "nav") {
            assert_array_equals(window.events, expectedEvents, 'incorrect event order');
            popup.close();
            resolve();
          }
      }));
    });
  }, `transition finishes with no named elements`);
} else if (is_popup_page) {
  onload = () => {
    requestAnimationFrame(() => requestAnimationFrame(() => {
      location.replace(location.href.split('?')[0] + '?new');
    }));

    onpageswap = (e) => {
      window.opener.events.push("pageswap");
    };

    onpagehide = () => {
      window.opener.events.push("pagehide");
    };
  };
} else if (is_new_page) {
  onpagereveal = (e) => {
    window.opener.events.push("pagereveal");
    e.viewTransition.finished.then(() => {
      window.opener.events.push("finished");
      channel.postMessage("nav");
    });
  }
}
</script>
